refactor geneset page
